move all bash_profile setup into bash_profile
I love how organized this was previously. However, all the need to link
files into specific places in order to have bash_profile be so nicely
organized is a real hassle when deploying my bash_profile to another
machine. In order to facilitate making that easier, I'm going with a
bash_profile that's all inclusive.